feat(tests): add deliberation tests and remove redundant test file
+ Added new tests for deliberation in ~deliberations.rs~
+ Removed redundant test file ~tests.rs~
+ Improved import formatting in ~main.rs~
