+++ packages/api/src/controllers/v2/organizations/_id/mod.rs
@@ -5,6 +5,3 @@ pub mod members;
 pub mod panels;
 pub mod resources;
 pub mod surveys;
-
-#[cfg(test)]
-mod tests;
