+++ packages/api/src/main.rs
@@ -9,14 +9,9 @@ use deliberation_resources::deliberation_resource::DeliberationResource;
 use discussions::Discussion;
 use models::step::Step;
 use models::{
-    deliberation::Deliberation,
-    deliberation_response::DeliberationResponse,
-    deliberation_user::DeliberationUser,
-    deliberation_vote::DeliberationVote,
-    invitation::Invitation,
-    response::SurveyResponse,
-    review::Review,
-    v2::{Institution, PublicOpinionProject},
+    deliberation::Deliberation, deliberation_response::DeliberationResponse,
+    deliberation_user::DeliberationUser, deliberation_vote::DeliberationVote,
+    invitation::Invitation, response::SurveyResponse, review::Review, v2::Institution,
 };
 use models::{organization::Organization, *};
 use sqlx::postgres::PgPoolOptions;
